Split batch controller test is two parts
